+++ packages/view/src/view-service-provider.ts
@@ -1,7 +1,7 @@
 
 import { ViewManager } from './view-manager.js'
+import { ViewResponse } from './view-response.js'
 import { ServiceProvider } from '@supercharge/support'
-import { View as ViewResponse } from './view-response.js'
 import { HttpResponse, HttpResponseCtor, ViewBuilderCallback } from '@supercharge/contracts'
 
 /**